feat: integrate profile data with backend and refine UI components

+++ lib/features/auth/presentation/providers/auth_provider.dart
@@ -65,8 +65,8 @@ class AuthController extends _$AuthController {
 
   Future<void> completeRegistration({
     required int age,
-    required double weight,
-    required double height,
+    required int weight,
+    required int height,
   }) async {
     state = const AuthState.loading();
     await Future.delayed(const Duration(milliseconds: 800));
